fix testcase
